Add files via upload
This GitHub repository hosts a machine learning project that aims to predict an individual's smoker status using bio-signals. Smoking is a significant public health concern, and early identification of smokers is crucial for targeted intervention and support. This project utilizes a dataset of bio-signals, such as heart rate, respiratory rate, and other physiological measurements, to develop a binary prediction model for classifying individuals as smokers or non-smokers.